박미란 - IdolController Session값추가

+++ test26/src/main/java/ksmart/project/test26/IdolController.java
@@ -2,6 +2,8 @@ package ksmart.project.test26;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,7 +21,10 @@ public class IdolController{
 	
 	// 목록조회
 	@RequestMapping(value="/idol/idolList",method=RequestMethod.GET)
-	public String IdolList(Model model) {
+	public String IdolList(Model model, HttpSession session) {
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect:/log/login";
+		}
 		List<Idol> list = idolDao.selectIdolList();
 		model.addAttribute("list",list);
 		return "idol/idolList";		
@@ -27,20 +32,29 @@ public class IdolController{
 	  
 	// 입력페이지 요청
 	@RequestMapping(value="/idol/idolInsert", method=RequestMethod.POST)
-	public String idolInsert(Idol idol) {
+	public String idolInsert(Idol idol, HttpSession session) {
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect:/log/login";
+		}
 		idolDao.insertIdol(idol);
 		return "redirect:/idol/idolList";
 	}
 	
 	// 입력 처리요청
 	@RequestMapping(value="/idol/idolInsert", method=RequestMethod.GET)
-	public String idolInsert() {
+	public String idolInsert(HttpSession session) {
+		if(session.getAttribute("loginMember")==null){
+			return "redirect:/log/login";
+		}
 		return "idol/idolInsert";
 	}	
 	
 	// 업데이트 정보요청
 	@RequestMapping(value="/idol/idolUpdate", method=RequestMethod.GET)
-	public String idolUpdate(Model model, @RequestParam(value="idolId", required=true) int idolId) {
+	public String idolUpdate(Model model, @RequestParam(value="idolId", required=true) int idolId, HttpSession session) {
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect:/log/login";
+		}
 		Idol idol = idolDao.updateGetIdol(idolId);
 		model.addAttribute("idol", idol);
 		return "idol/idolUpdate";
@@ -48,14 +62,20 @@ public class IdolController{
 	
 	// 업데이트 처리요청
 	@RequestMapping(value="/idol/idolUpdate", method=RequestMethod.POST)
-	public String idolUpdate(Idol idol) {
+	public String idolUpdate(Idol idol, HttpSession session) {
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect:/log/login";
+		}
 		idolDao.updateIdol(idol);
 		return "redirect:/idol/idolList";
 	}
 	
 	// 삭제
 	@RequestMapping(value="/idol/idolDelete" ,method=RequestMethod.GET)
-	public String idolDelete(@RequestParam(value="idolId", required=true) int idolId) {
+	public String idolDelete(@RequestParam(value="idolId", required=true) int idolId, HttpSession session) {
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect:/log/login";
+		}
 		idolDao.deleteIdol(idolId);
 		return "redirect:/idol/idolList";
 	}
